@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
 import { TemplateRole } from '../models/template-role';
 import { RoleService } from '../services/role.service';
+import { MatDialog, MatDialogConfig } from '@angular/material';
+import { RoleDetailComponent } from '../role-detail/role-detail.component';
 
 @Component({
   selector: 'app-role-list',
@@ -9,11 +11,18 @@ import { RoleService } from '../services/role.service';
 export class RoleListComponent implements OnInit {
   templateRoles: TemplateRole[];
 
-  constructor(private roleService: RoleService) {}
+  constructor(private roleService: RoleService, private dialog: MatDialog) {}
 
   ngOnInit() {
     this.roleService
       .getAllRoles()
       .subscribe(templateRoles => (this.templateRoles = templateRoles));
+  }
+
+  openDialog() {
+    const dialogConfig = new MatDialogConfig();
+    dialogConfig.disableClose = true;
+    dialogConfig.autoFocus = true;
+    this.dialog.open(RoleDetailComponent, dialogConfig);
   }
 }
